+++ src/model/calculation.h
@@ -27,7 +27,7 @@ class Calculation {
    * @brief Retrieves the result of calculation
    * @return double value representing the result of the calculation
    */
-  double GetResult();
+  long double GetResult();
 
   /**
    * @brief Iterates through the expression, pushing numbers into stack and
@@ -38,12 +38,12 @@ class Calculation {
    * @param expression - input expression to be calculated
    * @param x_value - value of x variable
    */
-  void Parse(std::string &expression, double x_value);
+  void Parse(std::string &expression, long double x_value);
 
  private:
   ReversePolishNotation rpn_;     ///< class converting input expression into RPN
   std::list<Lexeme> expression_;  ///< expression in RPN to calculate
-  double calculation_result_;     ///< result of calculations
+  long double calculation_result_;     ///< result of calculations
 
   /**
    * @brief Performs arithmetic and trigonometric calculations based on the
@@ -53,7 +53,7 @@ class Calculation {
    * @param numbers - stack of double values representing the operands
    * @return result of calculations as a double
    */
-  double Calculate(s21::Lexeme &current_operator, std::stack<double> &numbers);
+  long double Calculate(s21::Lexeme &current_operator, std::stack<long double> &numbers);
 
 };  // class Calculation
