+++ src/model/plot_graph.h
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <vector>
+
 #include "calculation.h"
 
 namespace s21 {
@@ -13,29 +14,27 @@ class PlotGraph {
 
   ~PlotGraph() = default;
 
-  std::vector<std::pair<double, double>> CalculateY(
-      std::string &expression, std::pair<double, double> x_range);
+  std::vector<std::pair<long double, long double>> CalculateY(
+      std::string &expression, std::pair<long double, long double> x_range);
 
  private:
-  double x_min_;
-  double x_max_;
-  std::vector<std::pair<double, double>> coordinates_;
+  long double x_min_;
+  long double x_max_;
+  std::vector<std::pair<long double, long double>> coordinates_;
   Calculation calc_expression_;
 
 };  // class PlotGraph
 
-std::vector<std::pair<double, double>> PlotGraph::CalculateY(
-    std::string &expression, std::pair<double, double> x_range) {
-
+std::vector<std::pair<long double, long double>> PlotGraph::CalculateY(
+    std::string &expression, std::pair<long double, long double> x_range) {
   x_min_ = x_range.first;
   x_max_ = x_range.second;
 
-  double y_coordinate = 0;
-  double x_coordinate = 0;
-  double step = 0.1;
+  long double y_coordinate = 0;
+  long double x_coordinate = 0;
+  long double step = 0.1;
 
-  for (double i = x_min_; i < x_max_; i += step) {
-
+  for (long double i = x_min_; i < x_max_; i += step) {
     calc_expression_.Parse(expression, i);
     y_coordinate = calc_expression_.GetResult();
     x_coordinate = i;
